Add endpoints to fetch all and by ID animes

+++ src/main/java/com/odeyalo/animerestapi/repository/AnimeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnimeRepository extends JpaRepository<Anime, Long> {
+public interface AnimeRepository extends JpaRepository<Anime, Integer> {
 }
